Cleanup, more options

+++ stalkerchan2.rb
@@ -5,6 +5,8 @@ require "nokogiri"
 require "open-uri"
 require "exifr"
 require "optparse"
+require "fileutils"
+
 
 class Image
 
@@ -25,11 +27,15 @@ class Image
       puts "Lade #{url}"
       filename = File.join(@folder, url[/\d+\..*/])     
       begin
-        File.open(filename, 'w') do |file|
-          file.write(open(url).read)           
-        end 
+        if !File.exists?(filename) then
+          File.open(filename, 'w') do |file|
+            file.write(open(url).read)           
+          end
+        else  
+          puts "File exists already: #{filename}" if File.exists?(filename)
+        end
         @@read_images[@href] = true
-      rescue OpenURI::HTTPError
+      rescue OpenURI::HTTPError => error
         puts error
         puts url + " already gone..."
         File.delete(filename) 
@@ -56,11 +62,15 @@ class Faden
     if @@read_threads[url.split("#")[0]] then
       puts "Thread schon gelesen!"
     else
-      @doc = Nokogiri(open(url))
-      @@read_threads[url] = true
-      @images = @doc/"a[@href*='files']"
-      @images.each do |element|
-        Image.new(@root,element[:href],@folder).fetch
+      begin
+        @doc = Nokogiri(open(url))
+        @@read_threads[url] = true
+        @images = @doc/"a[@href*='files']"
+        @images.each do |element|
+          Image.new(@root,element[:href],@folder).fetch
+        end
+      rescue OpenURI::HTTPError => error
+        puts "Error while fetching #{url} - #{error}"
       end
     end  
   end
@@ -68,8 +78,9 @@ end
 
 class Page
   
-  def initialize(root, suffix, channel, pagenum, folder)
-    @root,@suffix,@channel,@pagenum,@folder = root, suffix, channel, pagenum,folder
+  def initialize(options, pagenum)
+    @channel,@pagenum,@folder,@options = options[:channel],pagenum,File.join(options[:folder],options[:channel]), options
+    @root, @suffix = "http://krautchan.net",".html" if @options[:chan] == "krautchan"
   end
 
   def page
@@ -85,9 +96,8 @@ class Page
     @doc = Nokogiri(open(url))
 
     @threads = @doc/"a[@href*='thread-']"
-
+    FileUtils.makedirs(@folder)
     @threads.each do |thread|
-      puts "Thread: #{@pagenum}"
       Faden.new(@root,thread[:href],@folder).fetch
     end
   end
@@ -95,20 +105,24 @@ end
 
 class Downloader
 
-  def initialize(root, suffix, channel, folder, options)
-    @root, @suffix, @channel,@folder,@options = root, suffix, channel, folder,options
-    if File.exists?(File.join(folder, channel + "_downloaded")) then
-      puts channel + "_downloaded existiert"
+  def initialize(options)
+    @channel,@folder, @options = options[:channel], options[:folder], options
+    if File.exists?(File.join(@folder, @channel + "_downloaded")) then
+      puts @channel + "_downloaded existiert"
     end
+  end
+
+  def really_fetch(pagenum)
+    Page.new(@options,pagenum).fetch
   end
 
   def fetch
     threads = [] if @options[:threaded]
     (0..10).each do |pagenum|
       if @options[:threaded] then
-        threads << Thread.new { Page.new(@root, @suffix, @channel, pagenum,@folder).fetch}        
+        threads << Thread.new { really_fetch(pagenum) }        
       else
-        Page.new(@root, @suffix, @channel, pagenum,@folder).fetch
+        really_fetch(pagenum)
       end
     end
     threads.each do |thread| thread.join end if @options[:threaded]
@@ -138,6 +152,26 @@ optparse = OptionParser.new do |opts|
     options[:threaded] = true
   end
 
+  options[:channel] = "b"
+  opts.on("-c","--channel CHANNEL","set channel to scrape (default: b)") do |channel|
+    options[:channel] = channel
+  end
+
+  options[:folder] = "images"
+  opts.on("-o","--output FOLDER","set output folder (default: images)") do |folder|
+    options[:folder] = folder
+  end
+
+  options[:chan] = "krautchan"
+  opts.on("-f","--fourchan","scrape 4chan instead of Krautchan") do
+    options[:chan] = "4chan"
+  end
+
+  options[:gps] = false
+  opts.on("-g","--gps","look for GPS data") do
+    options[:gps] = true
+  end
+
   opts.on("-h","--help","Display this screen") do
     puts opts
     exit
@@ -146,4 +180,4 @@ end
 
 optparse.parse!
 
-Downloader.new("http://krautchan.net",".html","b","images", options).fetch
+Downloader.new(options).fetch
